Move the log page to yarn module.
